pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodal-compression-framework"
version = "0.1.0"
description = "A framework for compressing multimodal models like CLAP and AudioCLIP"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "transformers>=4.21.0",
    "datasets>=2.0.0",
    "optimum>=1.0.0",
    "onnx>=1.16.0",
    "onnxscript>=0.1.0",
    "onnxruntime>=1.12.0",
    "hydra-core>=1.3.0",
    "wandb>=0.13.0",
    "typer>=0.7.0",
    "streamlit>=1.20.0",
    "plotly>=5.10.0",
    "matplotlib>=3.6.0",
    "numpy>=1.21.0",
    "scipy>=1.9.0",
    "tqdm>=4.64.0",
]

[project.scripts]
multimodal-compress = "src.framework.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"